[endpoint-test] Reconfigure directory structure and rename to ProjectTester
