+++ src/main/scala/temple/test/internal/EndpointConfig.scala
@@ -1,3 +1,3 @@
-package temple.test
+package temple.test.internal
 
 private[test] case class EndpointConfig(baseIP: String, kongAdminURL: String)
